added showslide page

+++ app/Http/Controllers/GalleryController.php
@@ -6,15 +6,14 @@ use Illuminate\Http\Request;
 use App\Gallery;
 use App\Image;
 use Illuminate\Support\Facades\Auth;
+use File;
+use Validator;
+//use Image;
 //use Intervention\Image\Facades\Image;
 class GalleryController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function __construct(){
+    
+    public function __construct(){ // for login auth
         $this->middleware('auth');
     }
 
@@ -25,30 +24,43 @@ class GalleryController extends Controller
        return view('gallery.gallery',compact('galleries'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+  
+
     public function create()
     {
         //
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
+public function slideshow(Gallery $gallery){
+  // foreach($gallery->images as $gall){
+  //   dd($gall->file_name);
+  // }
+  
+
+   
+  // $albumname = $gallery->name;
+    return view('gallery.slideshow',compact('gallery'));
+   }
     public function store(Request $request)
     {
-        //dd($request->all());
-        $request->validate([
-            'name' => 'required|unique:galleries|min:5|max:100',
-        ],[ 'name.required' => 'Name of the gallary is required','name.unique'=>'This name is already taken, choose a unique name']);
 
-        Gallery::create($request->all());
+       $validate= $request->validate(
+            [
+                'name' => 'required|unique:galleries|min:5|max:100',
+                'image_path'=>'required|mimes:jpeg,png,jpg,gif|max:1028|dimensions:max_width=200,max_height=200',
+            ],
+            [   // for custom Validation Msg
+                'name.required' => 'Name of the gallary is required',
+                'name.unique'=>'This name is already taken, choose a unique name',
+                'image_path.dimensions'=>'Image dimensions should be 200 X 200 '
+            ]);
+
+        $photos = $request->file('image_path');
+        $save_name = 'gallery_img_'.$photos->getClientOriginalName();
+        $photos->move(public_path('galleryCoverImgs/'.$request->name), $save_name);
+        $path= '/galleryCoverImgs/'.$request->name.'/'.$save_name ;
+
+        Gallery::create(array_merge($validate,['image_path'=>$path]));
         return redirect('/gallery?page=1');
     }
 
@@ -62,69 +74,42 @@ class GalleryController extends Controller
     {
         //dd($gallery->id);
 
-        $images = Image::where('gallery_id',$gallery->id)->orderByRaw('order_img = 0, order_img ASC')->get();
+        $images = Image::where('gallery_id',$gallery->id)
+            ->orderByRaw('order_img = 0, order_img ASC')->get();
 
-        
-       
         return view('gallery.show',compact('gallery','images'));
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function edit($id)
     {
         //
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function update(Request $request, $id)
     {
-         $request->validate([
+         $request->validate(
+            [
             'order_img' => 'nullable|numeric',
-        ],['order_img.numeric'=>'Input Numeric value only']);
+            ],
+            ['order_img.numeric'=>'Input Numeric value only']);
+
         $image = Image::findOrFail($id);
-       $image->update(['order_img'=>$request->order_img]);
-       return redirect()->back();
+        $image->update(['order_img'=>$request->order_img]);
+        return redirect()->back();
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
 
     public function imageUpload(Request $request){
-        /*$file = $request->file;
-       
-        $filename = $request->file('file')->getClientOriginalName();
-        $file->move(public_path('galleryImg/images'), 'gh'); 
-        $gallery = Gallery::findOrFail($request->gallery_id);
-        $image = $gallery->images()->create([
-            'gallery_id'=>$request->gallery_id,
-            'file_name'=>$request->file->getClientOriginalName(),
-            'file_size'=>$file->getClientSize(),
-            'file_mime'=>$file->getClientMimeType(),
-            'file_path'=>'gallery/images' . $filename,
-            'created_by'=>Auth::user()->id,
-        ]);*/
 
         $gallery = Gallery::findOrFail($request->gallery_id);
-         $photos = $request->file('file');
-            $save_name = $photos->getClientOriginalName();
+        $photos = $request->file('file');
+        $save_name = $photos->getClientOriginalName();
        
-         $photos->move(public_path('galleryImg/'.$gallery->name), $save_name);
-         
+        $photos->move(public_path('galleryImg/'.$gallery->name), $save_name);
+        
 
          $image = $gallery->images()->create([
             'gallery_id'=>$request->gallery_id,
@@ -134,17 +119,50 @@ class GalleryController extends Controller
             'file_path'=>'/galleryImg/' . $gallery->name .'/'. $save_name,
             'created_by'=>Auth()->user()->id,
         ]);
+
+        return "uploaded";
         
         
     }
 
     public function destroy(Gallery $gallery)
     {
-         $gallery->delete();
-  
-        return redirect()->route('gallery.index')
-                        ->with('success','Product deleted successfully');
+         
+       
+
+foreach($gallery->images as $image){
+
+   $image_path = public_path().$image->file_path;
+
+   if(file_exists($image_path)) {
+    unlink($image_path);
+    
+    
+     }
+    
     }
+
+    $galleryImgPath = public_path().$gallery->image_path;
+    if(file_exists($galleryImgPath)) {
+    unlink($galleryImgPath);
+    
+     }
+
+    $gallery->images()->delete(); // Delete inside image table
+    $rem = public_path('/galleryImg/'.$gallery->name);
+    $remGalleryImg = public_path('/galleryCoverImgs/'.$gallery->name);
+    $gallery->delete();
+    if(is_dir($rem)) { // For exception handling Dir not found
+     rmdir($rem);
+    }
+    if(is_dir($remGalleryImg)) { // For exception handling Dir not found
+     rmdir( $remGalleryImg);
+    }
+
+    return  redirect()->back()->with('success', 'The Gallery Folder along with the associated images have been deleted Successfully! '); 
+   
+    }
+
     public function destroyImage(Image $image ,$id)
     {
